hardcoded line and bar charts
