+++ frontend/src/components/Dashboard/Dashboard.js
@@ -66,9 +66,9 @@ const Dashboard = (props) => {
       
       <Row xs={1} md={2} className={styles.cardRow}>
         <InsightCard title='Reddit' text='Reddit' isLoggedIn={redditSuccess} email={email} code={code} navigate={props.navigate} setExternalUrl={props.setExternalUrl}/>
-        <InsightCard title='Twitter' text='put Twitter stuff here' />
-        <InsightCard title='Instagram' text='put Instagram stuff here' />
-        <InsightCard title='YouTube' text='put YouTube stuff here' />
+        <InsightCard title='Twitter' text='put Twitter stuff here' isLoggedIn={true}/>
+        <InsightCard title='Instagram' text='put Instagram stuff here' isLoggedIn={true}/>
+        <InsightCard title='YouTube' text='put YouTube stuff here' isLoggedIn={true}/>
       </Row>
     </div>
   );
